+++ packages/ActionBar/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Removed <Filters> component from this package. Please use `@paprika/filters` instead.
+
 ## [0.2.0] - 2020-06-08
 
 ### Added